fix(types): clear the 11 standing mypy errors so the check can gate
`pixi run -e dev typecheck` was configured but not enforced, and sat at 11
errors across 5 files. A checker with known-permitted failures is one people
learn to skim: c5739d35 shipped `exc_info=True` on an rclpy logger, mypy
reports it as a plain call-arg error, and a 12th line in an 11-line list is
invisible. That bug then killed the navigator on the track (see 420216fb).

218 source files, zero errors.

Two of the eleven were real defects, not type noise:

  * BuildHatDriver.run_drive_forward/reverse narrowed `speed` to `int | None`
    against a DriveDriver base declaring `int | float | None` -- a Liskov
    violation, so a float speed through the base interface was unsound.
    Widened to `float | None`; `_clamp_speed` already coerced with `int()`.

  * SignSpec.color was typed `str` with a comment saying it holds SignColor,
    and was passed to `outward_lateral_axis(color: SignColor)`. Typing it
    truthfully immediately surfaced two more places leaking raw `str` into it
    -- ObservedSignTrack.color, and a `camera_color.value` in the router that
    should have been the enum. SignColor is a StrEnum, so equality and
    formatting are unchanged.

The rest were the type system unable to express a true invariant:

  * `corner_arc_radius` "too many arguments": joining two lambda types across
    a conditional loses the returned tuple's LENGTH, so `*effective(...)` read
    as arbitrary arity. Annotated the variable.

  * `lateral_offset: float | None` into `float` parameters, twice.
    __post_init__ resolves the None default before construction returns, so no
    instance carries None. One `_lateral_offset` property states that once,
    replacing three sites -- one of which already carried its own bare assert.

  * `EncoderConfig()` missing counts_per_rev/max_rpm, four times. Pydantic v2
    is dataclass_transform-ed, so mypy synthesises a strict __init__ for a
    settings class whose values arrive from TOML, and those two fields are
    deliberately REQUIRED with no default. `HardwareBaseSettings.load()`
    constructs through `cls`, which sidesteps the synthesis with no ignore
    needed anywhere -- warn_unused_ignores caught the one I tried to add.

  * Two `Driver` classes bound to one name by lazy conditional import.
    Aliased to Bts7960Driver / L298nDriver.

Also fixes the 4 ruff findings these changes introduced (PYI041, TC003). The
2 remaining ARG005 in generation.py pre-date this, verified by linting the
HEAD revision.

Verification: sign_discovery, sign_router, sign_lane, core_navigator_sign_lanes,
waypoints, encoder all pass -- the SignColor change is the only behaviour-
touching one and those are the suites that cover it. The 2 failures in
test_hardware_profile are byte-identical at HEAD with these files reverted, so
they are pre-existing and unrelated.

+++ platform/robot/src/hardware/motors/build_hat/driver.py
@@ -186,7 +186,7 @@ class Driver(MotorDriver):
         """Get current drive speed in degrees/s."""
         return float(self.drive.get_speed())
 
-    def _clamp_speed(self, speed: int) -> int:
+    def _clamp_speed(self, speed: float) -> int:
         """Clamp speed to configured limits."""
         speed = abs(speed)
         return int(clamp(speed, self.config.drive.min_speed, self.config.drive.max_speed))
@@ -196,14 +196,14 @@ class Driver(MotorDriver):
         return clamp(position, self.config.steering.left_limit_angle, self.config.steering.right_limit_angle)
 
     @override
-    def run_drive_forward(self, speed: int | None = None) -> None:
+    def run_drive_forward(self, speed: float | None = None) -> None:
         """Run drive motor forward with optional speed limit."""
         s = self._clamp_speed(speed or self.config.drive.default_speed)
         self.logger.info("Starting drive forward", extra={DETAILS_KEY: {"speed": s}})
         self.drive.start(s)
 
     @override
-    def run_drive_reverse(self, speed: int | None = None) -> None:
+    def run_drive_reverse(self, speed: float | None = None) -> None:
         """Run drive motor in reverse."""
         s = self._clamp_speed(speed or self.config.drive.default_speed)
         self.logger.info("Starting drive reverse", extra={DETAILS_KEY: {"speed": -s}})
